IndexedDB/tests/submissions/Opera/key_invalid.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>Invalid key</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script type="text/javascript">
    function invalid_key(desc, key) {
        var t = async_test(document.title + " - " + desc);

        createdb(t).onupgradeneeded = function(e) {
            var objStore = e.target.result.createObjectStore("store");

            assert_throws('DataError', function() {
                objStore.add("value", key);
            });
        };
    }

    invalid_key('null', null);
    invalid_key('NaN', NaN);
    invalid_key('function()', function() {});
    invalid_key('{ obj: 1 }', { obj: 1 });
    invalid_key('Math', Math );
    invalid_key('[{}]', [{}]);
    invalid_key('[[], [], [], [[ Date ]]]', [ [], [], [], [[ Date ]] ]);

    var sparse = [];
    sparse[10] = "hei";
    invalid_key('sparse array', sparse);

    var sparse2 = [];
    sparse2[0]  = 1;
    sparse2[""] = 2;
    sparse2[2]  = 3;
    invalid_key('sparse array', sparse2);

    invalid_key('[[1], [3], [7], [[ sparse array ]]]', [ [1], [3], [7], [[ sparse2 ]] ]);

//    var recursive = [];
//    recursive.push(recursive);
//    invalid_key('array directly contains self', recursive);

//    var recursive2 = [];
//    recursive2.push([recursive2]);
//    invalid_key('array indirectly contains self', recursive2);

//    var recursive3 = [recursive];
//    invalid_key('array member contains self', recursive3);
</script>
</script>
</script>

<div id="log"></div>
